Add files via upload
Building REST APIs with Python, Flask, Flask-RESTful:
Creating secure and reliable REST APIs which accepts text data from the client, processes it and stores it in a database, and then returns some data so the client can show it to the user which also includes authentication, logging, caching, and token refresh. Employing SQLAlchemy and Flask-SQLAlchemy to easily and efficiently store resources to a database.